requirements.txt
selenium>=4.0.0
webdriver-manager>=4.0.0
playwright>=1.40.0
python-dotenv>=1.0.0
requests>=2.28.0
